==> code/06_h2a_tract_intersections.ipynb <==
import pandas as pd
import geopandas as gpd
from bs4 import BeautifulSoup
from urllib.request import Request, urlopen
import requests
import re
from geopandas.tools import sjoin
import rtree
import plotnine as p9
from plotnine import *
import numpy as np

# get census tract shapes
 

main_url = 'https://www2.census.gov/geo/tiger/TIGER2014/TRACT/'

raw_content = requests.get(main_url).text
parsed_content = BeautifulSoup(raw_content, 'html.parser')

find_links = parsed_content.findAll("a", attrs = {"href": re.compile("tract")})
zip_names = [link.get('href') for link in find_links]
zip_links = [main_url + link for link in zip_names]

list_of_tract_shapes = [gpd.read_file(link) for link in zip_links]
tract_shapes = pd.concat(list_of_tract_shapes)
tract_shapes.head()

# get intersections of h2a violations and tracts


h2a = pd.read_pickle("../../qss20_finalproj_rawdata/summerwork/intermediate/h2a_combined_2014-2021_geocoded.pkl")
print(h2a.columns)
h2a_geom = gpd.GeoDataFrame(h2a, geometry=gpd.points_from_xy(h2a.geo_long, h2a.geo_lat))
h2a_geom.head()
h2a_geom

## filter out NA coordinates
nan_h2a_geom=h2a_geom.loc[pd.isnull(h2a_geom[['geo_lat', 'geo_long']]).any(axis=1)]
filtered_h2a_geom=h2a_geom.loc[~pd.isnull(h2a_geom[['geo_lat', 'geo_long']]).any(axis=1)]


h2a_tract_joined = sjoin(filtered_h2a_geom.set_crs(tract_shapes.crs), tract_shapes, how = 'left', op = 'intersects')

## rowbind the NA data to the merged data & Write pkl
final_tract_joined = h2a_tract_joined.append(nan_h2a_geom)
# final_tract_joined.shape
final_tract_joined.to_pickle("h2a_tract_intersections.pkl")

# map visual


cases_by_tract = final_tract_joined.groupby('GEOID').CASE_NUMBER.agg('nunique').reset_index()
merged = gpd.GeoDataFrame(cases_by_tract.merge(final_tract_joined[['GEOID', 'STATEFP', 'geometry']], 
                          how = 'right', on = 'GEOID').drop_duplicates(), geometry = 'geometry')
merged.sort_values('CASE_NUMBER')


plot = merged
plot['CASE_NUMBER'] = plot['CASE_NUMBER'].replace(0, np.nan)


p9.options.figure_size = (10, 8)

ggplot(plot) + \
    geom_map(aes(fill = 'CASE_NUMBER'), size = 0) + \
    scale_fill_gradient(low='yellow',high='red', na_value = 'grey') + \
    xlim(-130, -65) + \
    ylim(24, 50)

final_tract_joined[final_tract_joined.GEOID == '16083001100']
