==> code/20a_combine_clean_addendums.ipynb <==
# Imports 

## packages
import pandas as pd
import numpy as np
import os
from langdetect import detect, detect_langs
import os
from os import listdir

## repeated printouts
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

# Research Questions:
### In this project, I ask two questions:
- 1) How can the job addendums predict the case status (fully certified or denied) of an H2-A application (an application that employers file to hire foreign workers)?  
- 2) How can the job addendums predict whether the job addendums has been translated or not (in other word whether the job addendums contain non-English information)?  


# Why are these reseach questions important?
- Allow government (particularly Department of Labor) to more efficiently and effectively spot potential violations as H-2A workers' rights are oftentimes neglected. As they do not necessarily speak English and usually work seasonally, their rights are easily deprived by employers for low wage, poor living conditions and inhumane workloads.
- Provide insights on the composition of job addendums as these are reflective and crucial to understand the nature of the jobs

# Overview of what each notebook does
- 1) 20a_combine_clean_addendums.ipynb: combine 2020 and 2021 job addendums data and flag any non-English addendums
- 2) 20b_google_cloud_addendums_tranlation.ipynb: translate any non-English addendums
- 3) 20c_merging_addendumns_status.ipynb: merge the H-2A Job Certificate Data (include the outcome variables "CASE_STATUS" that we want to predict) with the job addendums data
- 4) 20d_preprocessing_modeling_1question.ipynb: pre processing job addendums, create a job postings classifier using TF-IDF and finally create and run a Multinomial Naive Bayes model to see the prediction of CASE_STATUS by the text features + explore the 
- 5) 20e_preprocessing_modeling_2question.ipynb: pre processing job addendums, create a job postings classifier using TF-IDF and finally create and run a Multinomial Naive Bayes model to see the prediction of translation status by the text features 


# Dataset used in this notebook
- read in: FOIA_2021-F-05932_raw_data.xlsx (raw data on each case's information and addendum that is requested by tThe Freedom of Information Act (FOIA) request) 
- export: addendum_wdetails_cleaned.csv (cleaned information and addendums data for cases that exclude cases with no addendums, NaN and undetectable addendums and include language detection result prepared for translation)

# Major tasks accomplished in this notebook
- 1) Combine job addendums sheets
- 2) Detect the job addendums' language

# Why are these tasks important
- By detecting the original language of each addendum, the translation process would be speed up (as it would waste a lot of time running over and translating Englush addendums). Moreover, by detecting non-English and excluding those that are English can allow me to ensure that all of my addendums get translated because free Google translation API only allows translation up to certain characters and the dataset that I am using is a big one.

# Directory

# dropbox
dropbox_general = "/Users/euniceliu/Dropbox (Dartmouth College)/"
DROPBOX_DATA_PATH = os.path.join(dropbox_general,
                                "qss20_finalproj_rawdata/summerwork/")
DATA_RAW_DIR = os.path.join(DROPBOX_DATA_PATH, "raw/")
DATA_ID_DIR = os.path.join(DROPBOX_DATA_PATH, "intermediate/")
WRITEFOLDER = os.path.join(DATA_ID_DIR)
# github
GITHUB_DATA_PATH = "../data/raw_data/"

#  Combine and clean addendums data 
#### Note: This section (combine and clean addendums data ) of the code was adopted and modified from the the start of the project written by Professor Johnson as we also combined the job addendums but it was across multiple years. In my project here, I limit to 2020 and 2021 year.

## read in multi-sheet excel file
all_foia_sheets = pd.read_excel(GITHUB_DATA_PATH + "FOIA_2021-F-05932_raw_data.xlsx",
                               sheet_name= None)

print("There are %s FOIA sheets"% len(all_foia_sheets))

## stored as dictionary with sheets as different keys; extract to inspect
fy20_addendums = all_foia_sheets['FOIA_2021-F-05932_FY2020']
fy21_q1_addendums = all_foia_sheets['FOIA_2021-F-05932_FY2021_Q1']

## confirm same columns
cols_20_not21 = set(fy20_addendums.columns).difference(fy21_q1_addendums.columns)
cols_21_not20 = set(fy21_q1_addendums.columns).difference(fy20_addendums.columns)

assert len(cols_20_not21) == 0
assert len(cols_21_not20) == 0 

## rowbind original dictionaries, drop fully duplicated rows, and remove index cols
addendum_init = pd.concat(all_foia_sheets).reset_index()
addendum_init2 = addendum_init.drop_duplicates(["CASE_NUMBER", "SECTION_DETAILS"]).copy()
addendum_init2['is_missing_all'] = (addendum_init2.SECTION_NAME.isnull()) & \
                                (addendum_init2.SECTION_NUMBER.isnull()) & \
                                (addendum_init2.SECTION_DETAILS.isnull())

## create filtered addendum df as: 
## (1) drop rows missing all details (later we'll do a left join to all disclosures) and 
## (2) remove levels columns from reset_index()
addendum = addendum_init2.loc[~addendum_init2.is_missing_all,
                             [col for col in addendum_init2.columns
                             if "levels" not in col]]

print("There are %s rows and %s unique job ids"% (addendum.shape[0], 
                                            len(addendum.CASE_NUMBER.unique())))

## Cleaning section name and section number
#### Note: This section (Cleaning section name and section number) of the code was adopted from the beginning of the project written by Professor Johnson who is also working on the project as we need to combine the job addendums
#### For now, leave unit of analysis as job-section dyad rather than concatenating / filtering to top sections at this phase


## explore relationship between section # and section name
addendum.SECTION_NAME.value_counts().head()
addendum.SECTION_NUMBER.value_counts().head()


## group by section # and see different section names
addendum.groupby('SECTION_NUMBER').agg({'SECTION_NAME': lambda x: "; ".join(sorted(x.astype(str).unique()))})

addendum.head()

# Detecting non-English language to figure out which ones to translate
## Major steps
- 1) detect language
- 2) separate dtected languages into different column

## Pre-Language Detection

### Generic function to detect language
#### Note: the below basic function detect_onestr was wrote by Professor Johnson to serve as a demo to detect different language in the job posting. However, the following sections of the code with more complex detection of language and testing are written by myself.

def detect_onestr(one_str):
    
    ## return list
    res = detect_langs(one_str)
    
    ## transform into a string and split on :
    split_res = str(res[0]).split(":")
    
    ## return split
    return(split_res)

### Test language detectionon on example with true positive in spanish CASE_NUMBER: H-300-20063-372516 and all english addendum CASE_NUMBER:H-300-19316-139384

sample_add = addendum.sample(n = 200, random_state = 91988)

## test language detection code on a couple examples
examples = sample_add.loc[sample_add.CASE_NUMBER.isin(["H-300-20063-372516",
                                                     "H-300-19316-139384"])].copy()
examples

## add language and probabilities to dataframe
examples['lang'] = [detect_onestr(one_str)[0] for one_str in examples.SECTION_DETAILS]

examples['lang_prob'] = [detect_onestr(one_str)[1] for one_str in examples.SECTION_DETAILS]

### More robust language detection function: I further improve and optimize the original languag detection function to allow detection of more than one language as I recognize that some of the job postings are combinations of multiple languages.

from langdetect import detect, DetectorFactory, detect_langs
import re

DetectorFactory.seed = 42

## first test the following strings has more than one language and 
## involving cases where there are 1-3 languages detected
detect_langs("hola, buenos dias! how are you")
detect_langs("muy bien. y tu? tengo hambre. i am good")
detect_langs("hola")

## functions that apply to the job postings
def detect_onestr(one_str):
    
    ## return list
    res = detect_langs(one_str)
    
    ## if the string has more than one language present
    if len(res)>1:
        split_res = str(res).split(":")
        split_res = str(split_res).split(",")
    
    else:
        ## transform into a string and split on :
        split_res = str(res).split(":")

    ## return split
    return(split_res)



### Clean NaN out from the job posting column (Section_Details) as detect_langs doesn't work on NaN

## Examine the NaN rows
print(addendum[addendum['SECTION_DETAILS'].isnull()])
## 11 NaN rows
len(addendum[addendum['SECTION_DETAILS'].isnull()].index)
## Data with 99615 rows originally with NaN
len(addendum.index)

## filter to non NaN one
addendum_wdetails = addendum[addendum['SECTION_DETAILS'].notna()]
# check the number of rows 99604 (correct! as 99615-11=99604)
len(addendum_wdetails.index)

## Language Detection

### Run langugae detection on all job postings
- I incorporated the except on nontdetectable langauge because detect_langs wont run on strings that do not have any information
- This step may take a while (around 10 minutes) to run as there are a lot of job postings. 


def detect_onestr(one_str):
    
    try:
        ## if the string has more than one language present
        ## return list
        res = detect_langs(one_str)
        if len(res)>1:
            split_res = str(res).split(":")
            split_res = str(split_res).split(",")

        else:
            ## transform into a string and split on :
            split_res = str(res).split(":")
    except:
        split_res="notdetectable"
    ## return split
    return(split_res)

## Take me about 10 minutes to run
addendum_wdetails['all']= [detect_onestr(one_str) for one_str in addendum_wdetails.SECTION_DETAILS]




## Post language detection cleaning up

### Remove notdectable job postings

## 66 rows' sectiondetails are not detectable
addendum_wdetails['all']=addendum_wdetails['all'].astype("string")
addendum_wdetails.loc[addendum_wdetails['all'] == "notdetectable"]
## filter out the notdetectable one
addendum_wdetails_cleaned=addendum_wdetails.loc[addendum_wdetails['all'] != "notdetectable"]
## Check the rows (99604-66=99538)
len(addendum_wdetails_cleaned.index)


### Test the above function on simple a pseudo multi-language string & create regrex to help separate the detection result in different columns

## create a pseudo dataframe with the above three strings
data = {'String': ['hola, buenos dias! how are you', 'muy bien. y tu? tengo hambre. i am good', "hola"]}  
df = pd.DataFrame(data)  

## Run and clean the column (to avoid bracket and '' )
## The pattern seems to work well seems to work well...
df['all']= [detect_onestr(one_str) for one_str in df.String]
df["lang1"] = df["all"].str[0].str.replace('[^\w\s]','')
df["lang1_prob"] = df["all"].str[1].str.replace('[^0-9,.]', '')
df["lang2"] = df["all"].str[2].str.replace('[^\w\s]','')
df["lang2_prob"] = df["all"].str[3].str.replace('[^0-9,.]', '')
df["lang3"] = df["all"].str[4].str.replace('[^\w\s]','')
df["lang3_prob"] = df["all"].str[5].str.replace('[^0-9,.]', '')
df

### Write more customized regrex and generate test case of real job addendums to check whether the splitting and regrex works well for job addendums 
- Here the regrex is a bit differnt from the above testing on the pseudo string because we need to customized way to separate different languages in job posting (such as removing white space). However, I decide to still keep the regrex testing on a pseudo non-English string because I build the above more complex one that work for job addendums one based on the regrex developed for simple non-English string.
- test on H-300-19277-071813, H-300-19277-071813 case


## Get the cases where there's more than 1 languages so we can check later after extraction
## H-300-19277-071813, H-300-19277-071813
addendum_wdetails_cleaned[addendum_wdetails_cleaned['all'].map(len) >2]

## test language detection code on a couple examples
test = addendum_wdetails_cleaned.loc[addendum_wdetails_cleaned.CASE_NUMBER.isin(["H-300-19277-071813",
                                                     "H-300-19277-071813"])].copy()
test.head()


## Clean out the punctuation
test["all"]=test["all"].str.replace('[^\w.]+',' ')

## replace more than one white space as one white space 
## (the reason why I am doing this is because if not then when extract
## the different values, the white space would count as one value)
test['all'] = test['all'].apply(lambda x: ' '.join(str(x).split()))

## Split by single white space so can ultimately extract three different lang
test["all"]=test["all"].str.split(" ")
test["lang1"] = test["all"].str[0]
test["lang1_prob"] = test["all"].str[1]
test["lang2"] = test["all"].str[2]
test["lang2_prob"] = test["all"].str[3]

## Seems to work well :)
test.head()


### Use splitting and regex to clean up the langauge detection result
- Apply the above tested regex and splitting to the all job postings that language have been detected
- There are 3 language columns (record up to 3 languages detecteed)
- Each langauge has a column and it's corresponding probability (lang_prob)
- If there's only 1 langauge the rest of the the language column will be NaN

# Clean out the punctuation
addendum_wdetails_cleaned["all"]=addendum_wdetails_cleaned["all"].str.replace('[^\w.]+',' ')
addendum_wdetails_cleaned.head()

## replace more than one white space as one white space 
## (the reason why I am doing this is because if not then when extract
## the different values, the white space would count as one value)
addendum_wdetails_cleaned['all'] = addendum_wdetails_cleaned['all'].apply(lambda x: ' '.join(str(x).split()))

## Split by white space so can ultimately extract three different lang
addendum_wdetails_cleaned["all"]=addendum_wdetails_cleaned["all"].str.split(" ")
addendum_wdetails_cleaned.head()

## Extract into different columns
addendum_wdetails_cleaned["lang1"] = addendum_wdetails_cleaned["all"].str[0]
addendum_wdetails_cleaned["lang1_prob"] = addendum_wdetails_cleaned["all"].str[1]
addendum_wdetails_cleaned["lang2"] = addendum_wdetails_cleaned["all"].str[2]
addendum_wdetails_cleaned["lang2_prob"] = addendum_wdetails_cleaned["all"].str[3]
addendum_wdetails_cleaned["lang3"] = addendum_wdetails_cleaned["all"].str[4]
addendum_wdetails_cleaned["lang3_prob"] = addendum_wdetails_cleaned["all"].str[5]
addendum_wdetails_cleaned.head()



### Check again on more than one language job postings!


addendum_wdetails_cleaned[addendum_wdetails_cleaned['all'].map(len) >2]

## Export the data that are ready for translation 


addendum_wdetails_cleaned.to_csv(WRITEFOLDER + 'addendum_wdetails_cleaned.csv')
