==> code/20b_google_cloud_addendums_translation.ipynb <==
import os
import pandas as pd
import numpy as np
import datetime
import yaml
import re

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import datetime
from datetime import datetime
from datetime import date

import nltk
from nltk import word_tokenize
nltk.download('punkt')
nltk.download('stopwords')

import os
from os import listdir


# Google Cloud Language Translation API
# We're using the basic version here == "v2" 
from google.cloud import translate_v2

import timeit



# Dataset used in this notebook
- read in: addendum_wdetails_cleaned.csv (cleaned information and addendums data for cases that exclude cases with no addendums, NaN and undetectable addendums and include language detection result prepared for translation)
- read in: sixth-topic-347322-4297c4a3d919.json (which is the Google Cloud Translation Credential) I did not include this file because the API has reached its free capacity for translation so even reading in the GOOGLE_APPLICATION_CREDENTIALS the translation would not run :( However, I thankfully did save the initial translation result.
- export: translated_job_postings (the job addendum data that continues all the translated addendum. in other words non-English addendum are all translated and presented in the dataset)

# Major tasks accomplished in this notebook
- 1) use the google cloud api to translate non-English addendums


# Why are these tasks important
- Instead of directly excluding the Non-English addendums, incorporating and translating those into English can make the study and the model more comprehensive and avoid potential biases.

# Directory

# dropbox
dropbox_general = "/Users/euniceliu/Dropbox (Dartmouth College)/"
DROPBOX_DATA_PATH = os.path.join(dropbox_general,
                                "qss20_finalproj_rawdata/summerwork/")
DATA_RAW_DIR = os.path.join(DROPBOX_DATA_PATH, "raw/")
DATA_ID_DIR = os.path.join(DROPBOX_DATA_PATH, "intermediate/")
WRITEFOLDER = os.path.join(DATA_ID_DIR)
# github
GITHUB_DATA_PATH = "../data/raw_data/"

## Read in Files & Data

## Google Cloud API Credential 
# os.environ['GOOGLE_APPLICATION_CREDENTIALS'] = r'/Users/euniceliu/Dropbox (Dartmouth College)/qss20_finalproj_rawdata/summerwork/raw/sixth-topic-347322-4297c4a3d919.json'


## Addendum with language identified
addendum_untranslated = pd.read_csv(DATA_ID_DIR + "addendum_wdetails_cleaned.csv")
                             

## Make Sure that Google Cloud API Work


# translate_client = translate_v2.Client()

# text = "hola como estas"
# target = "en"

# output = translate_client.translate(text, target_language=target)

# print(output)

## Function to Run Translation on Dataset
#### Note: This function below was adopted and adjusted from Professor Johnson's other project on education.  

# def google_translate_messages(one_row):
#     '''
#     Pass in a df row.
#     Find the message under the 'SECTION_DETAILS column'
#     '''
#     one_message = one_row.loc['SECTION_DETAILS']
    
#     # initialize the Google Cloud translation client
#     translate_client = translate_v2.Client()
    
#     # set the target language
#     target = 'en'
    
    
#     try:
#     # apply the translation 
#         output = translate_client.translate(one_message, 
#                                             format_='html',
#                                             target_language=target)
#     except:
#         output = {'translatedText': 'translation_error', 
#                   'detectedSourceLanguage': 'translation_error', 
#                   'orig_content':'translation_error'}

#     return list(output.values())

## Data Manipulation and Cleaning Before Translating

### Check if there is duplicated rows because google translation takes a lot of time


# it is good to prevent running translation twice
# len(addendum_untranslated)
# addendum_untranslated[['SECTION_DETAILS']].drop_duplicates()
# len(addendum_untranslated)


### Filter in non English and English rows



# nonEnglish = addendum_untranslated.loc[(addendum_untranslated['lang1']!="en") | (addendum_untranslated['lang2'].notnull()) | (addendum_untranslated['lang3'].notnull())]
# English = addendum_untranslated.loc[(addendum_untranslated['lang1']=="en") & (addendum_untranslated['lang2'].isnull()) & (addendum_untranslated['lang2'].isnull())]

## make sure the rows do add up
# assert len(addendum_untranslated)==len(nonEnglish)+len(English)

### Split the datast into multiple small dataframes (avoid code break and easier to debug as running the translating would take a while and we dont want to waste the Google Translation API capacity)
#### Note: This method/code is also adopted from Professor Johnson's other project on education.


# split_df = np.array_split(nonEnglish, 10)

# # N rows/columns per df. 
# for i in range(len(split_df)):
#     print('df', i+1, ':', 
#         split_df[i].shape)

## Set up to run translation
#### Note: This following three sections of code (before merging) is modified from Professor Johnson's other project on education.



# output path
# filename = 'translated_msgs_'
# ext = '.csv'


# # For every dataframe in split_df, 
# # run through the translation, unpack the results, 
# # and save as csv

# for i in range(len(split_df)):
#     one_df = split_df[i]
    
#     start_translation_time = timeit.default_timer() #time start
    
#     # run translation
#     one_df['output_list'] = one_df.apply(google_translate_messages, axis = 1)

#     stop_translation_time = timeit.default_timer() #time end
    
#     time_lapse = stop_translation_time - start_translation_time
#     print("took " + str(time_lapse) + " seconds to run")
    
#     # unpack the translation results into their own columns
#     one_df[['translatedText', 'detectedSourceLanguage', 'orig_content']] = \
#         pd.DataFrame(one_df.output_list.to_list(),  
#                      index = one_df.index)

#     one_df.to_csv(filename + str(i) + ext)
    
#     print("wrote results for df ", i + 1)

### Read all the translated files



## output path
filename = 'translated_msgs_'
ext = '.csv'

## init first df
translated_msgs_init = pd.read_csv(filename + '0' + ext)
print(translated_msgs_init.shape)
# initialize list to store our pickles
csv = []

## For every file in the folder, read it in as a dataframe, 
## then append to the csv list
## total 10 split datasets 
for i in range(1, len(split_df)):
    df = pd.read_csv( filename + str(i) + ext)
    print(df.shape)
    csv.append(df)

### Combine the 10 datasets together


translated_data = pd.concat([translated_msgs_init, *csv])
# Look at the languages detected
# translated_msgs.detectedSourceLanguage.value_counts()
# Check if the combined translated datasets match the number of rows of the original nonEnglish datasets
assert len(translated_data)==len(nonEnglish)

## Merge back with English (Untranslated) Job Postings

## Create an indicator for the addendums that are being translated
translated_data['translated'] = 'yes'
## Merge back the Job Posting with purely english
translated_postings = English.append(translated_data, ignore_index=True)
assert len(translated_postings)==len(addendum_untranslated)

translated_postings.head()

## Clean the columns 
## 1) replace translatedText and orig_content of the original english addendum with Section_Details
translated_postings.orig_content.fillna(translated_postings.SECTION_DETAILS, inplace=True)
translated_postings.translatedText.fillna(translated_postings.SECTION_DETAILS, inplace=True)
## 1) replace translated column that is NAN to "NO"
translated_postings.translated = translated_postings.translated.fillna('No')
## 2) get rid of irrelvant columns
translated_postings_cleaned = translated_postings.loc[:, ~translated_postings.columns.isin(['Unnamed: 0.1', 'output_list',"all","is_missing_all","detectedSourceLanguage"])]
translated_postings_cleaned.head()


### Write the translated dataset to csv file


translated_postings_cleaned.to_csv(DATA_ID_DIR+'translated_job_postings.csv')
