==> code/20d_preprocessing_modeling_1question.ipynb <==
import os
import pandas as pd
import numpy as np
import datetime
import yaml
import re
import time


from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import datetime
from datetime import datetime
from datetime import date

import nltk
from nltk import word_tokenize
nltk.download('punkt')
nltk.download('stopwords')


# Google Cloud Language Translation API
# We're using the basic version here == "v2" 
from google.cloud import translate_v2

import timeit

import os
from os import listdir


## nltk imports
# ! pip install gensim
from nltk.tokenize import word_tokenize, wordpunct_tokenize
from nltk.corpus import stopwords
from nltk.stem.porter import PorterStemmer
from gensim import corpora, models
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
import sklearn.metrics as metrics
from sklearn.feature_extraction.text import CountVectorizer




# Dataset used in this notebook
- read in: combined_certificate_postings.csv (the merged data that contains the job addendums and the CASE_STATUS variables)


# Major tasks accomplished in this notebook
- 1) Classify the outcome variables (CASE_STATUS) into binary 0,1
- 2) Pre-process the job addendums
- 3) Create a job posting classifier Using TF-IDF 
- 4) Create and run a Multinomial Naive Bayes model to see the prediction of CASE_STATUS by the text features
- 5) Compare the result of Multinomial Naive Bayes model that use TF-IDF score vs count
- 6) Explore the Bernoulli Naive Bayes model

# Why are these tasks important
- Provide insights on how the job addendums help predict the Case_status  
- Showcase the strength of incorporating TF-IDF score into my model which support that my model (creating a job posting classifier using TF-IDF scores and create and run a Multinomial Naive Bayes model) is the better choice

# Directory

# dropbox
dropbox_general = "/Users/euniceliu/Dropbox (Dartmouth College)/"
DROPBOX_DATA_PATH = os.path.join(dropbox_general,
                                "qss20_finalproj_rawdata/summerwork/")
DATA_RAW_DIR = os.path.join(DROPBOX_DATA_PATH, "raw/")
DATA_ID_DIR = os.path.join(DROPBOX_DATA_PATH, "intermediate/")
WRITEFOLDER = os.path.join(DATA_ID_DIR)
# github
GITHUB_DATA_PATH = "../data/raw_data/"

## Pre-Modeling

## Read in the combined dataset
combined_certificate_postings= pd.read_csv(DATA_ID_DIR+'combined_certificate_postings.csv')
combined_certificate_postings.head()

### Classify the outcome variables CASE_STATUS into 0 or 1 binary outcome 

## 1) check the category
combined_certificate_postings.CASE_STATUS.unique()
combined_certificate_postings['CASE_STATUS'].value_counts()


## 2) Classify the CASE_STATUS into binary variables (0 and 1)
## if certification +  certification (expired) >> 1 (approved cases)
## if partial certification + partial certification (expired) + denied (no approved cases)
combined_certificate_postings['CASE_OUTCOME'] = np.where(combined_certificate_postings['CASE_STATUS'].str.contains('Denied|Partial'), 0, 1)


## 3) check if coding to binary vairable work
combined_certificate_postings['CASE_OUTCOME'].unique()
combined_certificate_postings['CASE_STATUS'].value_counts()
combined_certificate_postings['CASE_OUTCOME'].value_counts()
# the number does add up!

###  Randomly Select 1000 Our Certified Case (Labelled as One)

## subset to certified case (label as 1) and not certified case (label as 0)
certified = combined_certificate_postings[combined_certificate_postings["CASE_OUTCOME"]==1]
notcertified = combined_certificate_postings[combined_certificate_postings["CASE_OUTCOME"]==0]
## show be 12977 rows
certified.shape
## show be 273 rows
notcertified.shape

## randomly select 1000 positive cases
certified_1000 = certified.sample(n = 1000)
certified_1000.shape
certified_1000.head()

## rowbind back to our notcertified dataset
combined_selected = pd.concat([certified_1000, notcertified])
combined_selected.shape

## Run Text Processing


### Convert characters to lower case

## lower case
start_time = time.time()
combined_selected['postings_lower']= combined_selected['combined_job_postings'].apply(lambda x: x.lower())
print("%s seconds" % (time.time() - start_time))


### Tokenize the words

## tokenized
start_time = time.time()
combined_selected['postings_tokenized'] = combined_selected['postings_lower'].apply(word_tokenize)
print("%s seconds" % (time.time() - start_time))

### Remove stopwords

## define stopwords
other_stopwords = ["after", "before", "employer", "employ", "job", "although", "provide", "complete","hour","time",
                  "begin","list","require","task","transportation","worker","workers","working","work","worked","works"]

list_stopwords = stopwords.words("english")+ other_stopwords

stopwords_complete = list_stopwords + other_stopwords
start_time = time.time()
## remove those
combined_selected['posting_without_stopwords']=combined_selected['postings_tokenized'].apply(lambda x: [word for word in x if word not in stopwords_complete])

print("%s seconds" % (time.time() - start_time))

### Perform Stemming

## stemming 
start_time = time.time()
porter = PorterStemmer()
combined_selected['stemmed'] = combined_selected['posting_without_stopwords'].apply(lambda x: [porter.stem(y)for y in x]) # Stem every word.
print("%s seconds" % (time.time() - start_time))

### Remove words that is less than 3 characters and punctuation

## keep isalpha() and the length of the word that is greater than 3
combined_selected['cleaned']=combined_selected['stemmed'].apply(lambda x: [word for word in x if word.isalpha() and len(word)>3])
combined_selected['cleaned']

### Join back each word


combined_selected['cleaned']=combined_selected['cleaned'].apply(lambda x: " ".join(x))
combined_selected['cleaned']

## Modeling 

### Split the data into training and testing

combined_selected.head()
combined_selected['CASE_OUTCOME'].value_counts()

from sklearn.model_selection import train_test_split
train,test=train_test_split(combined_selected,test_size=0.2)

train.head()

## Create a job postings classifier using TF-IDF

vectorizer = TfidfVectorizer(min_df=5, max_df=0.9)

### Write transform and fit into a function so the code looks neater

def tfidf_topwords(df,kind):
    if kind=="test":
        tf_idf = vectorizer.transform(df['cleaned'])
    elif kind=="train":
        tf_idf = vectorizer.fit_transform(df['cleaned'])
        tf_idf = vectorizer.transform(df['cleaned'])
    else:
        return ("wrong input")
    return tf_idf

X_train_tf=tfidf_topwords(train,"train")
X_test_tf=tfidf_topwords(test,"test")

X_train_tf.shape

X_test_tf.shape

## Create and run a Multinomial Naive Bayes model + Modeling Result 

# y value
train_y=train["CASE_OUTCOME"]
test_y=test["CASE_OUTCOME"]

naive_bayes_classifier = MultinomialNB()
naive_bayes_classifier.fit(X_train_tf, train_y)
y_pred = naive_bayes_classifier.predict(X_test_tf)
print(metrics.classification_report(test_y, y_pred, target_names=['Certified', 'Denied']))

print("Confusion matrix:")
print(metrics.confusion_matrix(test_y, y_pred))

#----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------#

## Findings using TF-IDF to create a job posting classifier & run a Multinomial Naive Bayes model


### As presented in the cells above, the data is extremely imbalanced (with 12977 positive cases (fully certified) and 273 negative cases (not fully certified & denied)). In this case, extra wrangling of the data is required or else the result of the predictive model would not be informative because the predictors will highly likely to predict any given test sample to the positive group which also means that the precision and recall will be very high. On the other hand, in this situation, the precision and recall would be very low for the negative cases. Thus, I randomly sampled the postive cases (case status that is fully certified) to only include 1000 cases to miminize the effect of imbalanced dataset. The model works better  as the accuracy is 0.8 and the weighted average for precision, recall and f1-score is relatively high which is respectively 0.81, 0.80 and 0.74.

# Extra exploration using the Bernoulli Naive Bayes model and Count Vectorizer + Multinomial Naive Bayes model

## Bernoulli Naive Bayes 

from sklearn.feature_extraction.text import CountVectorizer
CVectorizer=CountVectorizer(ngram_range=(1,3),min_df=5)
uni_bigrams=CVectorizer.fit(train['cleaned'])
X_train_vecotrized=uni_bigrams.transform(train['cleaned'])
X_test_vecotrized=CVectorizer.transform(test['cleaned'])


train_y=train["CASE_OUTCOME"]
test_y=test["CASE_OUTCOME"]

from sklearn.naive_bayes import BernoulliNB
BernoulliNB_classifier = BernoulliNB()
BernoulliNB_classifier.fit(X_train_vecotrized, train_y)
y_pred = BernoulliNB_classifier.predict(X_test_vecotrized)
print(metrics.classification_report(test_y, y_pred, target_names=['Certified', 'Denied']))


## Count Vectorizer + Multinomial Naive Bayes model

naive_bayes_classifier = MultinomialNB()
naive_bayes_classifier.fit(X_train_vecotrized, train_y)
y_pred = naive_bayes_classifier.predict(X_test_vecotrized)
print(metrics.classification_report(test_y, y_pred, target_names=['Certified', 'Denied']))

### After running the Bernoulli Naive Bayes model, I recognize that it would not make sense to make direct comparison of that with the Multinomial Naive Bayes model since I not only use different model but also the vectorizer is different. In this case, it is hard to analyze whether the better/worse performance of the model is due to the vectorizer choice or the model choice. Thus, I run another Multinomial Naive Bayes model but instead of using the TF-IDF score for words, I decided to use the count of the words for text classification because I am interested to know whether my initial choice of going for term frequency-inverse document frequency (TF-IDF) is the correct choice. We can see that the performace of model is better in the first model (the one using TF-IDF) as the accuracy score is 0.80 wheras in the second model (when using the count of the words) the accuracy score is 0.60. This makes sense because TF-IDF help to capture both the importance and relevance of a word whereas simply counting the occurance of words might include words that are too common across job postings which makes the prediction less accurate.


## Future Works

### Indeed, there might be some bias generated from only include the 1000 positive cases. I choice this way of dealing with the imbalanced data issue because it is the most intuitive and easiest way to quickly fix the issue. Future works can include performing other methods that tackle the imbalanced data issue such performing data augmentation, which can allow further verification of the results. Moreover, I also think it would be interesting see how having a addendum or not (as some of the cases do not have addendums--those cases are dropped in the very beggining) predict the case status.
